# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pi Control Hub",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions/register": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Register Session",
        "description": "Register a new session.",
        "operationId": "register_session_sessions_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Sessions Admin",
        "description": "List sessions for current user.",
        "operationId": "list_sessions_admin_sessions_get",
        "parameters": [
          {
            "name": "device_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  },
                  "title": "Response List Sessions Admin Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/telegram/link-token": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Link Token",
        "description": "Create a link token for Telegram binding.",
        "operationId": "create_link_token_telegram_link_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramLinkTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLinkTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TelegramLinkTokenRequest": {
        "properties": {
          "device_token": {
            "type": "string",
            "title": "Device Token"
          }
        },
        "type": "object",
        "required": [
          "device_token"
        ],
        "title": "TelegramLinkTokenRequest"
      },
      "TelegramLinkTokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "bot_username": {
            "type": "string",
            "title": "Bot Username"
          },
          "bot_link": {
            "type": "string",
            "title": "Bot Link"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "TelegramLinkTokenResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionRegisterRequest": {
        "properties": {
          "device_token": {
            "type": "string",
            "title": "Device Token"
          },
          "external_session_id": {
            "type": "string",
            "title": "External Session Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "project_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Path"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "running"
          }
        },
        "type": "object",
        "required": [
          "device_token",
          "external_session_id"
        ],
        "title": "SessionRegisterRequest"
      },
      "SessionRegisterResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status"
        ],
        "title": "SessionRegisterResponse"
      },
      "SessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "external_session_id": {
            "type": "string",
            "title": "External Session Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "project_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Path"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "substatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substatus"
          },
          "last_activity_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "external_session_id",
          "status",
          "created_at"
        ],
        "title": "SessionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}